Added and updated the capability of running cold-start and replay mode

+++ ush/cube2gaussian.sh
@@ -11,6 +11,13 @@ export REND=$FHMAX
 export rmhydro=${rmhydro:-".true."}
 export pseudo_ps=${pseudo_ps:-".true."}
 export phy_data=${phy_data:-""}
+export atminc="_atminc"
+
+if [[ $atminc = ".true." ]]; then
+  export rmhydro=".false."
+  export pseudo_ps=".false."
+  export phy_data=""
+fi
 
 if [[ $RHR == 0 ]]; then
   export fhour=$((DELTIM/3600.))
@@ -22,30 +29,39 @@ export diag_fhr=$((fhour+2*iau_halfdelthrs))
 export RDATE=$($NDATE +$RHR $sCDATE)
 
 COMOUT=${COMOUTatmos:-"."}
-$NLN $memdir/${APREFIX}logf$( printf "%03d" $fhour)_c2g $DATA/logf$( printf "%03d" $fhour)
+if [[ $atminc = ".true." ]]; then
+  AFHR=inc
+else
+  AFHR=f$( printf "%03d" $fhour)
+fi
+
+$NLN $memdir/${APREFIX}log${AFHR}_c2g $DATA/log${AFHR}
 
 GAUSSIANATMSSH=${GAUSSIANATMSSH:-$HOMEgfs/ush/gaussian_c2g_atms.sh}
 
 $GAUSSIANATMSSH
 
-ls $memdir/${APREFIX}atmf$( printf "%03d" $fhour)${ASUFFIX} > /dev/null 2>&1
+ls $memdir/${APREFIX}atm${AFHR}${ASUFFIX} > /dev/null 2>&1
 export err=$?
 $ERRSCRIPT||exit 2
 
 auxfhr=_auxfhr
-if [[ $auxfhr = "YES" ]]; then
-   GAUSSIANSFCSH=$HOMEgfs/ush/gaussian_sfcfcst_nodiagvar.sh
-else 
-   GAUSSIANSFCSH=$HOMEgfs/ush/gaussian_sfcfcst.sh
+
+if [[ $atminc = ".false." ]]; then
+  if [[ $auxfhr = "YES" ]]; then
+     GAUSSIANSFCSH=$HOMEgfs/ush/gaussian_sfcfcst_nodiagvar.sh
+  else 
+     GAUSSIANSFCSH=$HOMEgfs/ush/gaussian_sfcfcst.sh
+  fi
+     
+  $GAUSSIANSFCSH
+  
+  ls $memdir/${APREFIX}sfcf$( printf "%03d" $fhour)${ASUFFIX} > /dev/null 2>&1
+  export err=$?
+  $ERRSCRIPT||exit 2
 fi
-   
-$GAUSSIANSFCSH
 
-ls $memdir/${APREFIX}sfcf$( printf "%03d" $fhour)${ASUFFIX} > /dev/null 2>&1
-export err=$?
-$ERRSCRIPT||exit 2
-
-if [[ $err == 0 && -s $memdir/ ]]; then
+if [[ $err == 0 && -s $memdir/ && $atminc = ".false." ]]; then
    printf " completed fv3gfs fhour=%.*f %s" 3 $fhour $CDATE > $memdir/${APREFIX}logf$( printf "%03d" $fhour).txt
 fi
